+++ backend/models/__init__.py
@@ -1,0 +1,2 @@
+from .sensor import Sensor
+from .medicao import Medicao
